+++ code/drasil-lang/lib/Language/Drasil/Chunk/DefinedQuantity.hs
@@ -7,16 +7,17 @@ module Language.Drasil.Chunk.DefinedQuantity (
   dqd, dqdNoUnit, dqd',
   dqdQd, dqdWr, tempdqdWr') where
 
+import Drasil.Database.UID (HasUID(uid))
+import Drasil.Language.Concept (ConceptChunk, cw)
+
 import Language.Drasil.Symbol (HasSymbol(symbol), Symbol)
 import Language.Drasil.Classes (NamedIdea(term), Idea(getA), Concept, Express(..),
   Definition(defn), ConceptDomain(cdom), IsUnit, Quantity)
-import Language.Drasil.Chunk.Concept (ConceptChunk, cw)
 import Language.Drasil.Expr.Class (sy)
 import Language.Drasil.Chunk.UnitDefn (UnitDefn, unitWrapper,
   MayHaveUnit(getUnit))
 import Language.Drasil.Space (Space, HasSpace(..))
 import Language.Drasil.Stages (Stage)
-import Drasil.Database.UID (HasUID(uid))
 
 import Control.Lens ((^.), makeLenses, view)
