+++ code/drasil-lang/package.yaml
@@ -10,6 +10,7 @@ description: Please see the README on GitHub at <https://github.com/JacquesCaret
 language: Haskell2010
 default-extensions:
 - StrictData
+- ConstraintKinds
 
 extra-source-files: []
